requirements.txt
arxiv==2.4.0
fastapi==0.129.2
langchain==1.2.10
langchain_classic==1.0.1
langchain_community==0.4.1
langchain_core==1.2.14
langchain_huggingface==1.2.0
pydantic==2.12.5
python-dotenv==1.2.1
Requests==2.32.5
streamlit==1.54.0
uvicorn==0.41.0
duckduckgo-search
beautifulsoup4
tavily-python